+++ adapters/base/src/main/java/org/apache/cassandra/sidecar/adapters/base/CassandraFactory.java
@@ -45,8 +45,8 @@ public class CassandraFactory implements ICassandraFactory
     /**
      * Returns a new adapter for Cassandra 4.0 clusters.
      *
-     * @param session              the session to the Cassandra database
-     * @param jmxClient            the JMX client to connect to the Cassandra database
+     * @param session                     the session to the Cassandra database
+     * @param jmxClient                   the JMX client to connect to the Cassandra database
      * @param localNativeTransportAddress the address and port on which this instance is configured to listen
      * @return a new adapter for the 4.0 clusters
      */
