+++ client/src/main/java/org/apache/cassandra/sidecar/client/request/TokenRangeReplicasRequest.java
@@ -29,8 +29,7 @@ public class TokenRangeReplicasRequest extends DecodableRequest<TokenRangeReplic
 {
     /**
      * Constructs a new request to retrieve information by keyspace from token-range replicas endpoint
-     *
-     * @param keyspace the keyspace in Cassandra
+     * @param keyspace the keyspace for which the token range replicas will be retrieved.
      */
     public TokenRangeReplicasRequest(String keyspace)
     {
